Changed Bundle version to 3.4.0; Created BytesToImagePlusConvertor and ImagePlusToBytesConvertor classes for sharing the serialization code between the org.knime.python and org.knime.python2 extensionpoints implementations.
