+++ org.knime.knip.knimepython/META-INF/MANIFEST.MF
@@ -21,7 +21,7 @@ Require-Bundle: org.knime.base;bundle-version="2.11.0",
  imagej-common;bundle-version="0.12.1",
  scifio-bf-compat;bundle-version="1.11.0",
  org.knime.python.nodes;bundle-version="3.0.0",
- org.knime.python2;bundle-version="3.3.0"
+ org.knime.python2;bundle-version="3.4.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ClassPath: .
 Export-Package: org.knime.knip.knimepython
