+++ org.knime.knip.knimepython/src/org/knime/knip/knimepython/ImgPlusSerializer.java
@@ -7,6 +7,7 @@ import java.util.List;
 
 import org.knime.knip.base.data.img.ImgPlusValue;
 import org.knime.knip.io.ScifioGateway;
+import org.knime.knip.serialization.ImgPlusToBytesConvertor;
 import org.knime.python.typeextension.Serializer;
 import org.knime.python.typeextension.SerializerFactory;
 
@@ -28,36 +29,20 @@ import net.imagej.axis.Axes;
 import net.imagej.axis.CalibratedAxis;
 
 /**
- * Serializing ImgPlus instances to byte stream. Used format is .tif.
- *
- * TODO: Use a Scijava Service in the background such that one could replace the
- * way images are serialized/deserialized
+ * Serializer using the extension points in org.knime.python.
  * 
- * @author Christian Dietz (University of Konstanz)
- *
+ * @author Clemens von Schwerin, KNIME.com, Konstanz, Germany
  */
 @SuppressWarnings("rawtypes")
 public class ImgPlusSerializer extends SerializerFactory<ImgPlusValue> {
 
-	/**
-	 * ImgSaver to write ImgPlus to stream as tif
-	 */
-	private ImgSaver m_saver;
-
-	/**
-	 * SCIFIO config to read/write images
-	 */
-	private SCIFIOConfig m_scifioConfig;
-
+	private final ImgPlusToBytesConvertor m_convertor;
 	/**
 	 * Constructor
 	 */
 	public ImgPlusSerializer() {
 		super(ImgPlusValue.class);
-		m_saver = new ImgSaver(ScifioGateway.getSCIFIO().getContext());
-		m_scifioConfig = new SCIFIOConfig();
-		m_scifioConfig.groupableSetGroupFiles(false);
-		m_scifioConfig.imgOpenerSetComputeMinMax(false);
+		m_convertor = new ImgPlusToBytesConvertor();
 	}
 
 	@Override
@@ -65,107 +50,10 @@ public class ImgPlusSerializer extends SerializerFactory<ImgPlusValue> {
 
 		return new Serializer<ImgPlusValue<?>>() {
 
-			private final Writer m_writer;
-
-			{
-				try {
-					m_writer = ScifioGateway.format().getWriterByExtension(".tif");
-				} catch (FormatException e) {
-					throw new RuntimeException(e);
-				}
-			}
-
 			@Override
 			public byte[] serialize(final ImgPlusValue<?> value) throws IOException {
-				final ImgPlus<?> imgPlus = TypeUtils.converted(value.getImgPlus());
-
-				try {
-					final ByteArrayHandle handle = new ByteArrayHandle();
-					populateMeta(m_writer, imgPlus, m_scifioConfig, 0);
-					// HACK Corresponds to filename
-					m_writer.getMetadata().setDatasetName("");
-					m_writer.setDest(new RandomAccessOutputStream(handle), 0);
-
-					m_saver.saveImg(m_writer, imgPlus.getImg(), m_scifioConfig);
-
-					m_writer.close();
-
-					return handle.getBytes();
-				} catch (Exception e) {
-					e.printStackTrace();
-					throw new RuntimeException(
-							"Could not serialize image. Possible reasons: Unsupported image type, dimensionality of the image,...");
-				}
+					return m_convertor.serialize(value);
 			}
 		};
-	}
-
-	/**
-	 * This method is copied from SCIFIO
-	 * 
-	 * FIXME/TODO: Remove when method available in SCIFIO (see
-	 * https://github.com/scifio/scifio/issues/233)
-	 * 
-	 * Uses the provided {@link SCIFIOImgPlus} to populate the minimum metadata
-	 * fields necessary for writing.
-	 * 
-	 * @param imageIndex
-	 */
-	private void populateMeta(final Writer w, final ImgPlus<?> img, final SCIFIOConfig config, final int imageIndex)
-			throws FormatException, IOException, ImgIOException {
-
-		final Metadata meta = w.getFormat().createMetadata();
-
-		// Get format-specific metadata
-		Metadata imgMeta = ScifioGateway.getSCIFIO().imgUtil().makeSCIFIOImgPlus(img).getMetadata();
-
-		final List<ImageMetadata> imageMeta = new ArrayList<ImageMetadata>();
-
-		if (imgMeta == null) {
-			imgMeta = new DefaultMetadata();
-			imgMeta.createImageMetadata(1);
-			imageMeta.add(imgMeta.get(0));
-		} else {
-			for (int i = 0; i < imgMeta.getImageCount(); i++) {
-				imageMeta.add(new DefaultImageMetadata());
-			}
-		}
-
-		// Create Img-specific ImageMetadatas
-		final int pixelType = ScifioGateway.getSCIFIO().imgUtil().makeType(img.firstElement());
-
-		// TODO is there some way to consolidate this with the isCompressible
-		// method?
-		final CalibratedAxis[] axes = new CalibratedAxis[img.numDimensions()];
-		img.axes(axes);
-
-		final long[] axisLengths = new long[img.numDimensions()];
-		img.dimensions(axisLengths);
-
-		for (int i = 0; i < imageMeta.size(); i++) {
-			final ImageMetadata iMeta = imageMeta.get(i);
-			iMeta.populate(img.getName(), Arrays.asList(axes), axisLengths, pixelType, true, false, false, false, true);
-
-			// Adjust for RGB information
-			if (img.getCompositeChannelCount() > 1) {
-				if (config.imgSaverGetWriteRGB()) {
-					iMeta.setPlanarAxisCount(3);
-				}
-				iMeta.setAxisType(2, Axes.CHANNEL);
-				// Split Axes.CHANNEL if necessary
-				if (iMeta.getAxisLength(Axes.CHANNEL) > img.getCompositeChannelCount()) {
-					iMeta.addAxis(Axes.get("Channel-planes", false),
-							iMeta.getAxisLength(Axes.CHANNEL) / img.getCompositeChannelCount());
-					iMeta.setAxisLength(Axes.CHANNEL, img.getCompositeChannelCount());
-				}
-			}
-		}
-
-		// Translate to the output metadata
-		final Translator t = ScifioGateway.getSCIFIO().translator().findTranslator(imgMeta, meta, false);
-
-		t.translate(imgMeta, imageMeta, meta);
-
-		w.setMetadata(meta);
 	}
 }
